% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_filtering_adegenet.R
\name{apply_sequence_filtering}
\alias{apply_sequence_filtering}
\title{Apply a sequence of filterings over a gl object}
\usage{
apply_sequence_filtering(gl, filt_sequence)
}
\description{
filt_sequence named list (param = param_name, threshold: t, operator: op)
}
