% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_read_base.R
\name{convert_gt_to_dosage}
\alias{convert_gt_to_dosage}
\title{Given a list of genotype calls, get the dosage of each one}
\usage{
convert_gt_to_dosage(locus, alt_allele, ploidity = 2)
}
\arguments{
\item{locus}{List. A list of genotype calls, e.g., c("AG", "GG", "AA").}

\item{alt_allele}{String. The alternative allele, e.g., "A", "G".}

\item{ploidity}{Integer. The ploidity level of the organism, default is 2 (diploid).}
}
\value{
A list of integers, representing the allelic dosages of the alternative allele
        for each genotype call in the input list.
}
\description{
This function takes a list of genotype calls, an alternative allele, and the ploidity level
of the organism as input, and returns a list of allelic dosages corresponding to each
genotype call in the input list.
}
\examples{
genotype_calls <- c("AG", "GG", "AA")
convert_gt_to_dosage(genotype_calls, "A")  # Returns list(1, 0, 2)
convert_gt_to_dosage(genotype_calls, "G")  # Returns list(1, 2, 0)
convert_gt_to_dosage(c("AAA", "GGG"), "A", 3)  # Returns list(3, 0) (triploid)
convert_gt_to_dosage(c(NA, "AG"), "A")  # Returns list(NA, 1)
}
