% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_input_adegenet.R
\name{extract_dart_gt}
\alias{extract_dart_gt}
\title{extract_dart_gt}
\usage{
extract_dart_gt(
  data,
  sample.name.row = 2,
  first.data.row = 9,
  first.data.col = 6
)
}
\arguments{
\item{data}{Dataframe. Dataframe with Allelic dosage data}

\item{sample.name.row}{Integer. Row number which the sample names are (default = 2)}

\item{first.data.row}{Integer. Row number of the first genotype call}

\item{first.data.col}{Integer. Col number of the first genotype call}
}
\value{
Dataframe. Returns a Dataframe with the allelic dosage data converted
to integers and the column names as the samples.
}
\description{
extract_dart_gt
}
