% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_filtering_adegenet.R
\name{filter_gl}
\alias{filter_gl}
\title{Filter function}
\usage{
filter_gl(gl, parameter, threshold, comparison_operator)
}
\description{
This function generalizes the filtering functions using the parameter name
and comparing in locus or individuals using the provided comparision operator.
A list indicating the used thresold, if the filter was performed over individuals
or locus and the indices of elements that meet the comparision.
}
