% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_read_base.R
\name{genocall_to_allelic_dosage}
\alias{genocall_to_allelic_dosage}
\title{Genotype call to allelic dosage of alternative allele}
\usage{
genocall_to_allelic_dosage(genotype_call, alt_allele, ploidity = 2)
}
\arguments{
\item{genotype_call}{String. Genotype call, e.g., "AG", "AAA" (for triploid).}

\item{alt_allele}{String. Alternative allele, e.g., "A", "G".}

\item{ploidity}{Integer. Ploidity level of the organism, default is 2 (diploid).}
}
\value{
Integer. The allelic dosage of the given genotype call for the alternative allele.
}
\description{
This function takes a genotype call, an alternative allele, and the ploidity level
of the organism as input, and returns the allelic dosage of the alternative allele
in the genotype call.
}
\details{
The genotype call is expected to be a string of characters representing the alleles,
with each allele being a single character. For example, "AG" represents a diploid
genotype with one allele being "A" and the other being "G".

The allelic dosage is the count of the alternative allele in the genotype call.
For example, if the genotype call is "AG" and the alternative allele is "A", the
allelic dosage would be 1.

If the genotype call is missing (represented as NA or an empty string), the
function returns NA.
}
\examples{
genocall_to_allelic_dosage("AG", "A")  # Returns 1
genocall_to_allelic_dosage("GG", "A")  # Returns 0
genocall_to_allelic_dosage("AAA", "A", 3)  # Returns 3 (triploid)
genocall_to_allelic_dosage(NA, "A")  # Returns NA
}
