% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_imputation.R
\name{get_accuracy}
\alias{get_accuracy}
\title{Compare the genotype calls (gt) of a reference genlight object
against an imputed gl. Measure the accuracy of the imputation
process comparing the reference gt with imputed gt}
\usage{
get_accuracy(ref_gl, imp_gl)
}
\arguments{
\item{ref_gl}{Rreference genlight instance}

\item{imp_gl}{Imputed genlight instance}
}
\description{
Compare the genotype calls (gt) of a reference genlight object
against an imputed gl. Measure the accuracy of the imputation
process comparing the reference gt with imputed gt
}
