% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_get_metrics.R
\name{get_ind_missing}
\alias{get_ind_missing}
\title{Get Individual Missingness}
\usage{
get_ind_missing(gl)
}
\arguments{
\item{gl}{A genlight object.}
}
\value{
A numeric vector of length equal to the number of individuals (samples),
  containing the missing rate for each individual.
}
\description{
Compute the missing rate for each individual (sample) in a genlight object.
The missing rate is a value between 0 and 1, where 0 indicates no missing data
for that individual, and 1 indicates that all loci have missing data for that individual.
}
\examples{
data(example_genlight)
ind_miss <- get_ind_missing(example_genlight)
head(ind_miss)
}
