% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_get_metrics.R
\name{get_loc_missing}
\alias{get_loc_missing}
\title{Get Locus Missingness}
\usage{
get_loc_missing(gl)
}
\arguments{
\item{gl}{A genlight object.}
}
\value{
A numeric vector of length equal to the number of loci (markers),
  containing the missing rate for each locus.
}
\description{
Compute the missing rate for each locus (marker) in a genlight object.
The missing rate is a value between 0 and 1, where 0 indicates no missing data
for that locus, and 1 indicates that all samples have missing data for that locus.
}
\examples{
data(example_genlight)
loc_miss <- get_loc_missing(example_genlight)
head(loc_miss)
}
