% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_imputation.R
\name{i_freq_impute}
\alias{i_freq_impute}
\title{Imputation with allele frequency}
\usage{
i_freq_impute(q_frq, ploidity = 2)
}
\description{
Assuming a bi-allelic marker, using the observed allelic frequency for one allele
is sampled the genotype call for any ploidity level
}
