% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_imputation.R
\name{impute_beagle}
\alias{impute_beagle}
\title{Impute missing genotypes using Beagle}
\usage{
impute_beagle(
  gl,
  jre_path,
  beagle_path,
  memory = "Xmx1g",
  burnin = 3,
  iterations = 12,
  seed = -99999,
  nthreads = 16
)
}
\arguments{
\item{gl}{A genlight object of exclusively diploid ploidity level.}

\item{jre_path}{String. Path to the Java JRE directory.}

\item{beagle_path}{String. Path to the Beagle JAR file.}

\item{memory}{String. Allowed memory for JVM. Default is "Xmx1g".}

\item{burnin}{Integer. Number of burn-in iterations. Default is 3.}

\item{iterations}{Integer. Number of phasing iterations. Default is 12.}

\item{seed}{Integer. Seed for the random number generator. Default is -99999.}

\item{nthreads}{Integer. Number of threads to use. Default is 16.}
}
\value{
A genlight object with imputed genotypes.
}
\description{
This function converts a genlight object to a VCF file, runs Beagle for imputation using the specified JRE,
reads the imputed VCF back, and returns the imputed genlight object.
}
