% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_input_adegenet.R
\name{read_DArTSeq_SNP}
\alias{read_DArTSeq_SNP}
\title{read_DArTSeq_SNP}
\usage{
read_DArTSeq_SNP(path, snp_id, chr_name, pos_name)
}
\arguments{
\item{path}{Path to the DArTSeq CSV file.}

\item{snp_id}{Column name of the SNP ID data in the CSV file.}

\item{chr_name}{Column name of the chromosome where the SNP is located in the CSV file.}

\item{pos_name}{Column name of the physical position where the SNP is located in the CSV file.}
}
\value{
A genlight object with mapped SNPs.
}
\description{
This function reads a DArTSeq SNP dataset from a CSV file and converts it into a genlight object using
the DartR library reading functions.
}
\examples{
read_DArTSeq_SNP("path/to/dartseq/file.csv", snp_id = "SnpID", chr_name = "Chr", pos_name = "Position")
}
