% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_input_adegenet.R
\name{read_hapmap}
\alias{read_hapmap}
\title{read_hapmap}
\usage{
read_hapmap(path, ploidity = 2, sep = c("", "/", "|"))
}
\arguments{
\item{path}{Path to the hapmap file (uncompressed).}

\item{ploidity}{Integer, ploidity level of the organism (default: 2).}

\item{sep}{String, separator that divides the alleles (default: "").}
}
\value{
A genlight object.
}
\description{
This function reads a hapmap file (uncompressed) and converts it into a genlight object.
}
\examples{
read_hapmap("path/to/hapmap/file.txt")
read_hapmap("path/to/hapmap/file.txt", ploidity = 4)
read_hapmap("path/to/hapmap/file.txt", sep = "|")
}
