% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_input_adegenet.R
\name{read_one_row_dart_gt_data}
\alias{read_one_row_dart_gt_data}
\title{read_one_row_dart_gt_data}
\usage{
read_one_row_dart_gt_data(
  path,
  sample.name.row = 2,
  first.data.row = 9,
  first.data.col = 6,
  snp_id_col
)
}
\arguments{
\item{path}{}

\item{sample.name.row}{}

\item{first.data.row}{}

\item{first.data.col}{}

\item{snp_id_col}{}
}
\description{
This function takes as input a path a DArT file where each variant is coded
in a single line. It splits the metadata and the allelic dosages.
}
