% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_tabular_geno}
\alias{read_tabular_geno}
\title{Read a text table and use common na representations for convert to NA}
\usage{
read_tabular_geno(path, sep = c("\\t", ","))
}
\arguments{
\item{path}{Path to tabular data}

\item{sep}{Delimitator of the table cells}
}
\value{
Returns a dataframe
}
\description{
Read a text table and use common na representations for convert to NA
}
