% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_input_adegenet.R
\name{read_vcf}
\alias{read_vcf}
\title{read_vcf}
\usage{
read_vcf(path, ploidity = 2, na_reps = c("-", "./."), sep = "/")
}
\arguments{
\item{path}{String. Path to the VCF file. It could be compressed.}

\item{ploidity}{Integer. Ploidity level of the organism. (Default = 2)}

\item{na_reps}{Vector. A vector containing the NA representations of genotype calls (default: empty).}
}
\value{
A genlight object.
}
\description{
This function reads a VCF file (compressed or uncompressed) and converts it into a genlight object.
}
\examples{
fl = "https://github.com/Breeding-Analytics/cgiarGenomics/raw/main/tests/vcf_fmt/diploid.vcf.gz"
tempfl <- tempfile(pattern = 'diploid', fileext = '.vcf.gz')
download.file(fl, destfile = tempfl)
dat.dose.vcf = read_vcf(tempfl, ploidity = 2)
print(dat.dose.vcf)
plot(dat.dose.vcf)
}
