% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_read_base.R
\name{replace_strings_with_integers}
\alias{replace_strings_with_integers}
\title{Replace a list of strings with their corresponding integer values}
\usage{
replace_strings_with_integers(lookup_table, strings_to_replace)
}
\arguments{
\item{lookup_table}{A named list or vector with string keys and integer values.}

\item{strings_to_replace}{A list of strings to be replaced with their corresponding integer values.}
}
\value{
A list of integers, where each string in the input list has been replaced with its
        corresponding integer value from the lookup table, or NA if no match was found.
}
\description{
This function takes two inputs: a named list or vector with string keys and integer values,
and a list of strings to be replaced. It replaces each string in the second list with the
corresponding integer value from the first list, based on the string keys.
}
\details{
If a string in the second list does not have a corresponding key in the first list,
it will be replaced with NA.
}
\examples{
lookup <- c(A = 1, B = 2, C = 3)
strings <- c("B", "A", "D", "C")
replace_strings_with_integers(lookup, strings)  # Returns list(2, 1, NA, 3)
}
