% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_imputation.R
\name{set_fraction_na}
\alias{set_fraction_na}
\title{Mask a fraction of genotype calls for measure the 
accuracy of an imputation method}
\usage{
set_fraction_na(gl, fraction = 0.1)
}
\arguments{
\item{gl}{A genlight instance}

\item{fraction}{percentage of data to mask [0 - 1], default = 0.1}
}
\value{
A genlight instance where the genotype matrix have as NA
the given fraction of missing data
}
\description{
Mask a fraction of genotype calls for measure the 
accuracy of an imputation method
}
