% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_vcf}
\alias{write_vcf}
\title{Write genlight object to VCF file}
\usage{
write_vcf(gl, output_path, na_rep = "./.")
}
\arguments{
\item{gl}{A genlight object.}

\item{output_path}{String. Path to the output VCF file. It could end in `.gz` for compression.}

\item{na_rep}{String. NA representation for genotypes in the VCF file. Default is "./.".}
}
\description{
This function converts a genlight object to a VCF file and writes it to an output path.
}
